perf(p2p): :zap: debounce `p2p_order_list` call and check if p2p is available before making the call
